Refactor authentication.js
